requirements.txt
PyYAML==6.0
matplotlib==3.5.1
numpy==1.22.2
